GUI enhancements
(1) nparts_density is no longer used in the interface. This was quite confusing.
(2) During txt export, an empty dialog was opened when the file to be written to existed. This dialog now shows a message informing the user that overwrite is not possible at the moment. This needs more implementation work!